\name{Insurance}
\docType{data}
\alias{Insurance}
\title{Insurance consumption across Italian provinces, 1998-2002}
\description{
  A panel of 103 observations

 \emph{number of observations} :  515

 \emph{observation} :  provinces

   \emph{country} :  Italy

}
\usage{data(Insurance)}
\format{A dataframe containing :
  \describe{
    \item{code}{the province code according to Istat}
    \item{year}{the year of observation}
    \item{ppcd}{real per capita premiums in 2000 euros, non-life insurance excluding mandatory motor third-party liability}
    \item{rgdp}{real per-capita GDP}
    \item{bank}{real per-capita bank deposits}
    \item{den}{population density per square Km}
    \item{rirs}{real interest rate on lending to families and small enterprises}
    \item{agen}{density of insurance agencies per 1000 inhabitants}
    \item{school}{share of people with second grade schooling or more}
    \item{vaagr}{share of value added, agriculture}
    \item{fam}{average number of family members}
    \item{inef}{judicial inefficiency index: average years to settle first degree of civil case}
    \item{trust}{survey result to the question "do you trust others?"}
    \item{dXX}{year dummies}
    \item{NorthWest}{macroregional dummy}
    \item{NorthEast}{macroregional dummy}
    \item{Centre}{macroregional dummy}
    \item{South}{macroregional dummy}
    \item{Islands}{macroregional dummy (Sicily and Sardinia)}
 }
}
\source{
     Giovanni Millo and Gaetano Carmeci, (2011) \dQuote{Non-life insurance consumption in Italy: a sub-regional panel data analysis}, \emph{Journal of Geographical Systems}, \bold{13:273--298}.

}
\author{Giovanni Millo}
\keyword{datasets}
