\name{itaww}
\docType{data}
\alias{itaww}
\title{Spatial weights matrix - Italian provinces}

\description{
Spatial weights matrix of the 103 Italian provinces as in 1992-2005.  

}

\usage{
data(itaww)
}

\format{
A matrix with elements different from zero if province i and j are
neighbors. Weights are row-standardized. Messina and Reggio Calabria, divided by the Messina Strait, are considered neighbours.}

\author{Giovanni Millo}

\keyword{datasets}
