\name{splm-package}
\alias{splm-package}
\alias{splm}
\docType{package}
\title{
Spatial panel models: estimation and testing
}
\description{
A comprehensive toolset for ML and GM estimation and diagnostic testing of econometric models for spatial panel data.
}
\details{
\tabular{ll}{
Package: \tab splm\cr
Type: \tab Package\cr
Version: \tab 0.0\cr
Date: \tab 2009-07-29\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
Put here: An overview of how to use the package, including the most important functions.
}
\author{
Giovanni Millo and Gianfranco Piras

Maintainer: Giovanni Millo <giovanni.millo@generali.com>

}
\references{
Here: Literature or other references for background information
}
\keyword{ package }
\seealso{
}
\examples{
##Here put simple examples of the most important functions
}
