\name{usaww}
\docType{data}
\alias{usaww}
\title{Spatial weights matrix - US states}

\description{
Spatial weights matrix of the 48 continental US States based on the queen contiguity criterium.  

}

\usage{
data(usaww)
}

\format{
A matrix with elements different from zero if state i and j are
neighbors. Weights are row standardized. According to the queen
contiguity criterium, Arizona and Colorado are considered neighbours.}

\author{Giovanni Millo}

\keyword{datasets}
